admin.py in people-app updated
