+++ acm_web_site/apps/business/people/forms.py
@@ -1,0 +1,20 @@
+from django import forms
+
+from .models import Member, Team
+
+
+class MemberAdminForm(forms.ModelForm):
+    picture = forms.FileField()
+
+    class Meta:
+        model = Member
+        fields = '__all__'
+
+
+class TeamAdminForm(forms.ModelForm):
+    picture = forms.FileField()
+
+    class Meta:
+        model = Team
+        fields = '__all__'
+
