+++ acm_web_site/apps/business/web_site/admin.py
@@ -7,9 +7,7 @@ from .forms import (
 	GalleryAdminForm,
 	AwardAdminForm,
 	ActivityAdminForm,
-	MemberAdminForm,
 	ProjectAdminForm,
-	TeamAdminForm,
 	TutorialAdminForm
 )
 
@@ -50,18 +48,6 @@ class ActivityAdmin(admin.ModelAdmin):
 		super().save_model(request, obj, form, change)
 
 
-class MemberAdmin(admin.ModelAdmin):
-	form = MemberAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['picture']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.picture = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
-
 class ProjectAdmin(admin.ModelAdmin):
 	form = ProjectAdminForm
 
@@ -73,17 +59,6 @@ class ProjectAdmin(admin.ModelAdmin):
 		obj.poster = uploaded_file_url
 		super().save_model(request, obj, form, change)
 
-
-class TeamAdmin(admin.ModelAdmin):
-	form = TeamAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['picture']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.picture = uploaded_file_url
-		super().save_model(request, obj, form, change)
 
 
 class TutorialAdmin(admin.ModelAdmin):
@@ -99,8 +74,6 @@ class TutorialAdmin(admin.ModelAdmin):
 
 
 admin.site.register(Contest)
-admin.site.register(Member, MemberAdmin)
-admin.site.register(Team, TeamAdmin)
 admin.site.register(Award, AwardAdmin)
 admin.site.register(Gallery, GalleryAdmin)
 admin.site.register(Tutorial, TutorialAdmin)
